Refactor feedback-related classes to update User entity import path for consistency with new package structure and enhance FeedbackController with project service integration
